Update Products and models

+++ src/main/java/de/hoomit/stockmanagement/service/ProductService.java
@@ -4,16 +4,15 @@ import de.hoomit.stockmanagement.domain.Product;
 import de.hoomit.stockmanagement.exception.ProductChangedException;
 import de.hoomit.stockmanagement.repository.ProductRepository;
 import jakarta.annotation.Resource;
+import java.util.Date;
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.orm.ObjectOptimisticLockingFailureException;
 import org.springframework.stereotype.Service;
-
-import java.util.Date;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
 
 @Service
 public class ProductService {
@@ -50,42 +49,44 @@ public class ProductService {
      * @param amount number of stock that must be added to current stock
      * @return an {@link Optional} containing the newly changed Product
      */
-    public Optional<Product> refillProductStock(final Long id, final long amount) {
+    public Optional<Product> refillProductStock(final Long id, final int amount) {
         LOGGER.info("Refilling Product Stock for Id: " + id);
 
-        return productRepository.findById(id).map(
-                product -> {
-                    final long newStock = product.getStock() + amount;
+        return productRepository
+            .findById(id)
+            .map(product -> {
+                final int newStock = product.getStock() + amount;
 
-                    LOGGER.info(String.format("Refill Product Stock with amount: %d, new Stock: %d", amount, newStock));
+                LOGGER.info(String.format("Refill Product Stock with amount: %d, new Stock: %d", amount, newStock));
 
-                    product.setStock(newStock);
-                    productRepository.save(product);
+                product.setStock(newStock);
+                productRepository.save(product);
 
-                    return product;
-                }
-        );
+                return product;
+            });
     }
 
-    public Optional<Product> decreaseProductStock(final Long id, final long amount) {
-        return productRepository.findById(id).map(product -> {
-            final long productStock = product.getStock();
-            if (amount > productStock) { // Do we have enough Products to Buy?
-                LOGGER.error(String.format("Exceeding the product stock of: %d", productStock));
+    public Optional<Product> decreaseProductStock(final Long id, final int amount) {
+        return productRepository
+            .findById(id)
+            .map(product -> {
+                final int productStock = product.getStock();
+                if (amount > productStock) { // Do we have enough Products to Buy?
+                    LOGGER.error(String.format("Exceeding the product stock of: %d", productStock));
 
-                throw new IllegalArgumentException("Not enough Product stock for Id: " + id);
-            }
+                    throw new IllegalArgumentException("Not enough Product stock for Id: " + id);
+                }
 
-            // Substract amount from current product stock
-            final long newStock = productStock - amount;
+                // Substract amount from current product stock
+                final int newStock = productStock - amount;
 
-            LOGGER.info(String.format("Decreased from Product Id: %s: %d Products, new Stock: %d", product.getId(), amount, newStock));
+                LOGGER.info(String.format("Decreased from Product Id: %s: %d Products, new Stock: %d", product.getId(), amount, newStock));
 
-            product.setStock(newStock);
-            productRepository.save(product);
+                product.setStock(newStock);
+                productRepository.save(product);
 
-            return product;
-        });
+                return product;
+            });
     }
 
     /**
@@ -97,7 +98,7 @@ public class ProductService {
      * @param amount number of stock that must be subtracted from current stock
      * @return an {@link Optional} containing the newly changed Product
      */
-    public Optional<Product> buyProduct(final Long id, final long amount) {
+    public Optional<Product> buyProduct(final Long id, final int amount) {
         LOGGER.info("Buying Product Id: " + id);
 
         return decreaseProductStock(id, amount);
@@ -106,16 +107,18 @@ public class ProductService {
     public Product updateProduct(final Product product) {
         LOGGER.info("Updating Product: " + product);
 
-        productRepository.findById(product.getId()).map(savedProduct -> {
-            savedProduct.setName(product.getName());
-            savedProduct.setStock(product.getStock());
+        productRepository
+            .findById(product.getId())
+            .map(savedProduct -> {
+                savedProduct.setName(product.getName());
+                savedProduct.setStock(product.getStock());
 
-            try {
-                return productRepository.save(savedProduct);
-            } catch (ObjectOptimisticLockingFailureException e) {
-                throw new ProductChangedException(savedProduct.getId());
-            }
-        });
+                try {
+                    return productRepository.save(savedProduct);
+                } catch (ObjectOptimisticLockingFailureException e) {
+                    throw new ProductChangedException(savedProduct.getId());
+                }
+            });
 
         return product;
     }
